first unittest added using Mockito
