+++ resources/smpp/ra/pom.xml
@@ -1,35 +1,48 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
+    <modelVersion>4.0.0</modelVersion>
 
-	<parent>
-		<groupId>org.restcomm.slee.resource.smpp</groupId>
-		<artifactId>smpp-server-ra</artifactId>
-		<version>7.0.0-SNAPSHOT</version>
-	</parent>
+    <parent>
+        <groupId>org.restcomm.slee.resource.smpp</groupId>
+        <artifactId>smpp-server-ra</artifactId>
+        <version>7.0.0-SNAPSHOT</version>
+    </parent>
 
-	<artifactId>smpp-server-ra-ra</artifactId>
-	<name>Core : SLEE : SMPP Server RA</name>
+    <artifactId>smpp-server-ra-ra</artifactId>
+    <name>Core : SLEE : SMPP Server RA</name>
 
-	<dependencies>
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>smpp-server-ra-ratype</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>smpp-server-ra-library</artifactId>
-			<version>${project.version}</version>
-		</dependency>	
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>smpp-server-ra-events</artifactId>
-			<version>${project.version}</version>
-		</dependency>			
-		<dependency>
-			<groupId>org.restcomm.smpp</groupId>
-			<artifactId>ch-smpp</artifactId>
-			<version>${ch.smpp.version}</version>
-		</dependency>			
-	</dependencies>	
+    <dependencies>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>smpp-server-ra-ratype</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>smpp-server-ra-library</artifactId>
+            <version>${project.version}</version>
+        </dependency>	
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>smpp-server-ra-events</artifactId>
+            <version>${project.version}</version>
+        </dependency>			
+        <dependency>
+            <groupId>org.restcomm.smpp</groupId>
+            <artifactId>ch-smpp</artifactId>
+            <version>${ch.smpp.version}</version>
+        </dependency>		
+                
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-all</artifactId>
+            <version>1.9.5</version>
+            <scope>test</scope>
+        </dependency>     
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+            <version>1.5.2</version>
+            <scope>test</scope>
+        </dependency>                           	
+    </dependencies>	
 </project>
